requirements.txt
pandas
matplotlib
streamlit
google-generativeai
tqdm
